+++ app/blog/layout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {BadBlogStateProvider} from "../../components/blog/BadBlogButton";
-import {UDFontStateProvider} from "../../components/blog/UDFontButton";
+import {UDFontStateProvider} from "../../components/blog/UDFontBlock";
 
 export default function BlogLayout({children}: { children: React.ReactNode }) {
   return (
