+++ components/GoogleFonts.tsx
@@ -20,7 +20,7 @@ const fontObjectToString = ({family, weight}: Font) => {
 const GoogleFonts = () => (
   <>
     <link rel="preconnect" href="https://fonts.googleapis.com"/>
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="true"/>
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="anonymous"/>
     <link
       href={
         'https://fonts.googleapis.com/css2?' +
