--- lib/blog/fileWatch.ts
@@ -1,9 +1,0 @@
-// This file has only empty function, but is needed for HMR with Markdown updates.
-// To enable HMR, the empty function must be called at a valid position due to optimization of Next.js
-// Please see next.config.js
-
-export const doMarkdownHMR = () => {
-}
-
- console.log()
- console.log()
