+++ next.config.js
@@ -15,7 +15,6 @@ const config = {
 
   experimental: {
     appDir: true,
-    typedRoutes: true,
     mdxRs: true,
   },
